ima_evm_setup: don't mention critical_data
May not be helpful so don't confuse users with it.

Bruno Meneguele explains:

This rule is used for measuring internal kernel data (ie. other kernel
subsystem's buffers), so there isn't a real use in userspace, but might be
critical for defining what data structures should be considered in the chain
of trust.